backend/requirements.txt
flask>=3.1
flask-cors>=5.0
openai>=1.68
supabase>=2.13
python-dotenv>=1.1
tiktoken>=0.8
PyPDF2>=3.0
python-docx>=1.1
openpyxl>=3.1
requests>=2.32
celery[redis]>=5.4
redis>=5.0
